+++ ttn-stack/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## [1.2.0] - 2024-03-08
+
+### Added
+
+- Support for autoprovision device and bucket separately
+- Support for configuring a preexisting bucket for device write callback
+ 
+### Security
+
+- Updated axios to version 1.6.7
+- Updated express to version 4.18.2
+- Updated vm2 to version 3.9.19
+
 ## [1.1.0] - 2024-02-06
 
 ### Added
@@ -19,6 +32,7 @@
 
 - Make sure Thinger server is running before starting service
 
-[1.1.1]: https://github.com/thinger-io/plugins/compare/ttn-stack/v1.0.4...ttn-stack/v1.1.0
-[1.4.0]: https://github.com/thinger-io/plugins/compare/ttn-stack/v1.0.3...ttn-stack/v1.0.4
-[1.3.0]: https://github.com/thinger-io/plugins/tree/ttn-stack/v1.0.3
+[1.2.0]: https://github.com/thinger-io/plugins/compare/ttn-stack/v1.2.0...ttn-stack/v1.2.0
+[1.1.0]: https://github.com/thinger-io/plugins/compare/ttn-stack/v1.0.4...ttn-stack/v1.1.0
+[1.0.4]: https://github.com/thinger-io/plugins/compare/ttn-stack/v1.0.3...ttn-stack/v1.0.4
+[1.0.3]: https://github.com/thinger-io/plugins/tree/ttn-stack/v1.0.3